courseapi/requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
Django==6.0
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-js-asset==3.1.2
django-oauth-toolkit==3.1.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==25.0
pillow==12.1.0
pycparser==2.23
PyMySQL==1.1.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
